+++ demo/js/indexControl.js
@@ -87,7 +87,7 @@ $("#btn-minimal-mfalogin-next").click(function() {
                 }
                 $("#minimal-mfalogin-next")
                     .html('<table><tbody><tr>' +
-                        '<td>Google Authenticator Code: </td><td><input type="text" id="minimal-mfalogin-googlecode"></td>' +
+                        '<td>Authenticator Code: </td><td><input type="text" id="minimal-mfalogin-googlecode"></td>' +
                         '</tr></tbody></table>' + 
                         '<button id="btn-minimal-mfalogin-login">Login</button>');
                 multiFactorAuthToken = ret.data.SecondFactorAuthentication.SecondFactorAuthenticationToken;